Return full response in get method for puppet_https
Prior to this commit, the get method in puppet_https only returned the
repsonse body instead of the full response from net/https. This is
inconsistant with how the other http methods work in this library. This
commit changes that by returning the full response instead of just the
body.