(#5) Ensure request was properly made before finding group id
Prior to this commit, if a classifier api request failed in the
get_group_id function, it would not do any checks to ensure that it was
successful and would attempt to find the id. This commit fixes that by
first ensuring that the request was successful before attempting to find
the group id.